+++ NOTI_Teacher/NOTI_Teacher/Screen/Home/ViewController/HomeVC+TableView.swift
@@ -7,40 +7,6 @@
 
 import UIKit
 
-// MARK: - UITableViewDataSource
-
-extension HomeVC: UITableViewDataSource {
-    func numberOfSections(in tableView: UITableView) -> Int {
-        viewModel.output.lessons.count
-    }
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        let section = viewModel.output.lessons[section]
-        
-        return section.isOpened ?? false
-        ? section.homeworks.count + 1 : 1
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let sectionCell = tableView.dequeueReusableCell(withIdentifier: LessonTVC.className) as? LessonTVC,
-              let homeworkTVC = tableView.dequeueReusableCell(withIdentifier: HomeworkTVC.className) as? HomeworkTVC,
-              let studentListTVC = tableView.dequeueReusableCell(withIdentifier: StudentTVC.className) as? StudentTVC
-        else { fatalError() }
-        let totalRows = tableView.numberOfRows(inSection: indexPath.section)
-        
-        switch indexPath.row {
-        case 0:
-            sectionCell.configureCell(viewModel.output.lessons[indexPath.section])
-            return sectionCell
-        case totalRows - 1:
-            return studentListTVC
-        default:
-            homeworkTVC.configureCell()
-            return homeworkTVC
-        }
-    }
-}
-
 // MARK: - UITableViewDelegate
 
 extension HomeVC: UITableViewDelegate {
